.pkg-build.otfVRA/backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.28.1
requests==2.32.5
sqlalchemy[asyncio]==2.0.48
asyncpg==0.30.0
python-multipart==0.0.22
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
prometheus-client==0.24.1
croniter==6.0.0
psutil==7.2.2
fpdf==1.7.2
pandas==2.3.2
openpyxl==3.1.5
PyYAML==6.0.3
setuptools==82.0.1
# ldap3 is optional — only needed when LDAP/AD authentication is configured.
# Install manually: pip install ldap3>=2.9.1
# or add the wheel to vendor/wheels/ for offline builds.
cryptography>=42.0.0
# Enterprise hardening additions
slowapi==0.1.9          # Rate limiting (brute-force protection)
authlib==1.3.2          # OIDC/OAuth2 SSO (Okta, Azure AD, Google, Keycloak)
itsdangerous==2.2.0     # Secure token signing for OIDC state param
